backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-pubsub==2.18.4
google-generativeai==0.3.1
supabase==1.0.3
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
